requirements.txt
streamlit>=1.30
python-dotenv>=1.0.1
pandas>=2.0
gspread>=6.0.0
google-auth>=2.0
python-dateutil>=2.9